+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Minor INVISIBILITY improvements. [#3760](https://github.com/dmdorman/hero6e-foundryvtt/issues/3760)
 - Bulky modifier effect is now removed when not present in current HDC upload. [#3738](https://github.com/dmdorman/hero6e-foundryvtt/issues/3738)
 - Copy/Paste tokens with "Append Incrementing Number" property now ensure unique numbers [#3759](https://github.com/dmdorman/hero6e-foundryvtt/issues/3759)
+- Fix brace maneuver [#3791](https://github.com/dmdorman/hero6e-foundryvtt/issues/3791)
+- Fix set maneuver's bonus [#3795](https://github.com/dmdorman/hero6e-foundryvtt/issues/3795)
 
 ### Version 4.2.17 20260221